+++ lib-multisrc/greenshit/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("lib-multisrc")
+}
+
+baseVersionCode = 1
+
+dependencies {
+    api(project(":lib:i18n"))
+}
