+++ lib-multisrc/greenshit/res/mipmap-hdpi/ic_launcher.png
