+++ src/pt/aurorascan/build.gradle
@@ -1,0 +1,10 @@
+ext {
+    extName = 'Aurora Scan'
+    extClass = '.AuroraScan'
+    themePkg = 'greenshit'
+    baseUrl = 'https://aurorascan.net'
+    overrideVersionCode = 0
+    isNsfw = false
+}
+
+apply from: "$rootDir/common.gradle"
