+++ src/pt/maidscan/build.gradle
@@ -1,9 +1,9 @@
 ext {
     extName = 'Maid Scan'
     extClass = '.MaidScan'
-    themePkg = 'madara'
-    baseUrl = 'https://empreguetes.site'
-    overrideVersionCode = 2
+    themePkg = 'greenshit'
+    baseUrl = 'https://novo.empreguetes.site'
+    overrideVersionCode = 45
     isNsfw = true
 }
